+++ Boss2D/platform/boss_platform_blank.cpp
@@ -983,7 +983,7 @@
         ////////////////////////////////////////////////////////////////////////////////
         // GRAPHICS
         ////////////////////////////////////////////////////////////////////////////////
-        void Platform::Graphics::SetScissor(float x, float y, float w, float h)
+        void Platform::Graphics::SetScissor(double x, double y, double w, double h)
         {
             BOSS_ASSERT("호출시점이 적절하지 않습니다", ViewAPI::CurPainter());
 
@@ -1025,7 +1025,7 @@
             BOSS_ASSERT("This is blank platform. You can use BOSS_PLATFORM_XXX.", false);
         }
 
-        void Platform::Graphics::SetZoom(float zoom, OrientationRole orientation)
+        void Platform::Graphics::SetZoom(double zoom, OrientationRole orientation)
         {
             BOSS_ASSERT("호출시점이 적절하지 않습니다", ViewAPI::CurPainter());
