+++ Boss2D/platform/boss_platform_native.cpp
@@ -1020,7 +1020,7 @@
         ////////////////////////////////////////////////////////////////////////////////
         // GRAPHICS
         ////////////////////////////////////////////////////////////////////////////////
-        void Platform::Graphics::SetScissor(float x, float y, float w, float h)
+        void Platform::Graphics::SetScissor(double x, double y, double w, double h)
         {
             BOSS_ASSERT("호출시점이 적절하지 않습니다", ViewAPI::CurPainter());
 
@@ -1063,7 +1063,7 @@
             BOSS_ASSERT("Further development is needed.", false);
         }
 
-        void Platform::Graphics::SetZoom(float zoom, OrientationRole orientation)
+        void Platform::Graphics::SetZoom(double zoom, OrientationRole orientation)
         {
             BOSS_ASSERT("호출시점이 적절하지 않습니다", ViewAPI::CurPainter());
             ViewAPI::CurPainter()->setZoom(zoom);
